move InterlockingGenerator to separate file
